+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Fix new mesh part not linked to its collection after mesh separate
 - Remove logging of meaningless ReferenceError messages
 - Fix error when setting datablock reference after initial synchronization (e.g. TextureNodeImage.image, ArrayModifier.object_offset, Scene.camera)
+- Add exception type when logging an exception
 
 
 # 0.17.0 (2020-10-15)